+++ packages/vault-server/src/index.ts
@@ -1,0 +1,10 @@
+export { createVaultServer, type VaultServer } from './handler';
+export { listVaultFiles, safeVaultPath, vaultRelative, isWatchedRelativePath } from './files';
+export {
+  createDebouncedEmitter,
+  retryUntilReadable,
+  startVaultWatcher,
+  toVaultRelativePath,
+  type ChangeKind,
+  type VaultChangeEvent,
+} from './watcher';